Use writev() instead of sendmsg() if flags is 0
writev() on Linux only copies the io vector to validate it; sendmsg()
requires a memory allocation in order to do so.

+++ src/lib/lwan-io-wrappers.h
@@ -25,7 +25,7 @@
 #include "lwan.h"
 
 ssize_t lwan_writev(struct lwan_request *request, struct iovec *iov,
-                    size_t iovcnt);
+                    int iovcnt);
 ssize_t lwan_send(struct lwan_request *request, const void *buf, size_t count,
                   int flags);
 void lwan_sendfile(struct lwan_request *request, int in_fd,
