+++ src/lib/lwan-response.c
@@ -459,7 +459,7 @@ bool lwan_response_set_event_stream(struct lwan_request *request,
 void lwan_response_send_event(struct lwan_request *request, const char *event)
 {
     struct iovec vec[6];
-    size_t last = 0;
+    int last = 0;
 
     if (!(request->flags & RESPONSE_SENT_HEADERS)) {
         if (UNLIKELY(!lwan_response_set_event_stream(request, HTTP_OK)))
